+++ projects/storefrontlib/src/cms-components/navigation/index.ts
@@ -1,2 +1,4 @@
 export * from './breadcrumb/breadcrumb.component';
 export * from './breadcrumb/breadcrumb.module';
+export * from './footer-navigation/footer-navigation.component';
+export * from './footer-navigation/footer-navigation.module';
