+++ projects/storefrontlib/src/lib/cms-lib/index.ts
@@ -1,8 +1,6 @@
 export * from './category-navigation/category-navigation.component';
 export * from './category-navigation/category-navigation.module';
 export * from './cms-lib.module';
-export * from './footer-navigation/footer-navigation.component';
-export * from './footer-navigation/footer-navigation.module';
 export * from './link/link.component';
 export * from './link/link.module';
 export * from './navigation/navigation.component';
